refactor: 컨트롤러 이름 변경
